# doc/api/openapi.yaml
openapi: 3.0.3
info:
  title: StudyFlow API (MVP P0)
  version: 1.0.0
  description: |
    基于 PRD（MVP P0）生成的 StudyFlow OpenAPI 规范。
    统一响应结构：{ code, message, data }。
servers:
  - url: http://localhost:3000
    description: Local Dev
...
paths:
  /projects:
    post:
      tags: [Projects]
      summary: 创建学习项目
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: 获取项目详情
      operationId: getProjectDetail
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
    put:
      tags: [Projects]
      summary: 更新项目
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
    delete:
      tags: [Projects]
      summary: 删除项目
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
      required: [code, message]
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: boolean
              example: true
...
    Project:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        goal:
          type: string
          nullable: true
        examDate:
          type: string
          format: date-time
          nullable: true
        dailyTimeBudget:
          type: integer
          nullable: true
        mode:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, createdAt, updatedAt]
    ProjectCreateRequest:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          nullable: true
        examDate:
          type: string
          format: date-time
          nullable: true
        dailyTimeBudget:
          type: integer
          nullable: true
        mode:
          type: string
          nullable: true
      required: [name]
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          nullable: true
        examDate:
          type: string
          format: date-time
          nullable: true
        dailyTimeBudget:
          type: integer
          nullable: true
        mode:
          type: string
          nullable: true
    ProjectResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Project'
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            progress:
              type: object
              nullable: true
            taskStats:
              type: object
              nullable: true
    ProjectDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ProjectDetail'
...